+++ src/components/map/o-map-events.interface.ts
@@ -6,4 +6,20 @@ export interface OMapEvents {
   onMove: EventEmitter<any>;
   onMoveEnd: EventEmitter<any>;
   onZoomLevelsChange: EventEmitter<any>;
+
+  onDrawEvent: EventEmitter<any>;
+
+  onDrawCreated: EventEmitter<any>;
+  onDrawEdited: EventEmitter<any>;
+  onDrawDeleted: EventEmitter<any>;
+  onDrawDrawStart: EventEmitter<any>;
+  onDrawDrawstop: EventEmitter<any>;
+  onDrawvertex: EventEmitter<any>;
+  onDrawEditStart: EventEmitter<any>;
+  onDrawEditMove: EventEmitter<any>;
+  onDrawEditResize: EventEmitter<any>;
+  onDrawEditvertex: EventEmitter<any>;
+  onDrawEditStop: EventEmitter<any>;
+  onDrawDeleteStart: EventEmitter<any>;
+  onDrawDeleteStop: EventEmitter<any>;
 }