+++ src/components/map/o-map-w-events.class.ts
@@ -1,16 +1,33 @@
 import { EventEmitter } from '@angular/core';
-// import * as L from 'leaflet';
+import { Observable } from 'rxjs/Observable';
 import { OMapEvents } from './o-map-events.interface';
 import { OMapBase } from './o-map.class';
+import { OMapComponent } from './o-map.component';
+import { IMapDrawControlEvent } from '../map-draw-controls/o-map-draw-controls-events.class';
 
-export class OMapWEvents extends OMapBase {
-  public events: OMapEvents = {
-    onClick: new EventEmitter<any>(),
-    onDrag: new EventEmitter<any>(),
-    onMove: new EventEmitter<any>(),
-    onMoveEnd: new EventEmitter<any>(),
-    onZoomLevelsChange: new EventEmitter<any>()
-  };
+export class OMapWEvents extends OMapBase implements OMapEvents {
+
+  onClick = new EventEmitter<any>();
+  onDrag = new EventEmitter<any>();
+  onMove = new EventEmitter<any>();
+  onMoveEnd = new EventEmitter<any>();
+  onZoomLevelsChange = new EventEmitter<any>();
+
+  onDrawEvent = new EventEmitter<any>();
+
+  onDrawCreated = new EventEmitter<any>();
+  onDrawEdited = new EventEmitter<any>();
+  onDrawDeleted = new EventEmitter<any>();
+  onDrawDrawStart = new EventEmitter<any>();
+  onDrawDrawstop = new EventEmitter<any>();
+  onDrawvertex = new EventEmitter<any>();
+  onDrawEditStart = new EventEmitter<any>();
+  onDrawEditMove = new EventEmitter<any>();
+  onDrawEditResize = new EventEmitter<any>();
+  onDrawEditvertex = new EventEmitter<any>();
+  onDrawEditStop = new EventEmitter<any>();
+  onDrawDeleteStart = new EventEmitter<any>();
+  onDrawDeleteStop = new EventEmitter<any>();
 
   constructor() {
     super();
@@ -22,9 +39,23 @@ export class OMapWEvents extends OMapBase {
       ['Click', 'Drag', 'Move', 'MoveEnd', 'ZoomLevelsChange'].forEach(eventName => {
         this.getMapService().map.on(
           eventName.toLowerCase(),
-          (evt) => this.events['on' + eventName].emit(evt)
+          (evt) => {
+            this['on' + eventName].emit(evt);
+          }
         );
       });
+
     });
+
+    this.onMapReady().subscribe((oMap: OMapComponent) => {
+      const observable: Observable<IMapDrawControlEvent> = oMap.getDrawControlEventsObservable();
+      if (observable) {
+        observable.subscribe((args: IMapDrawControlEvent) => {
+          this.onDrawEvent.emit(args.data);
+          this[args.event].emit(args.data);
+        });
+      }
+    });
+
   }
 }