+++ src/components/map/o-map.component.ts
@@ -7,9 +7,9 @@ import { MapService, GeocodingService, TranslateMapService } from '../../service
 import { Util } from '../../utils';
 import { OMapWSearch } from './o-map-w-search.class';
 import * as L from 'leaflet';
+import { OMapCrsComponent } from '../map-crs/o-map-crs.component';
 
 //TODO import {Control} from 'leaflet-draw';
-
 
 const DEFAULT_INPUTS = [
   'sAttr: attr',
@@ -28,17 +28,39 @@ const DEFAULT_INPUTS = [
   'sBaseLayerIds: base-layer-ids'
 ];
 
+const DEFAULT_OUTPUTS = [
+  'onToggleWSLayerSelected',
+  'onToggleWSLayerVisibility',
+  'onToggleWSLayerInWS',
+
+  'onClick',
+  'onDrag',
+  'onMove',
+  'onMoveEnd',
+  'onZoomLevelsChange',
+
+  'onDrawEvent',
+
+  'onDrawCreated',
+  'onDrawEdited',
+  'onDrawDeleted',
+  'onDrawDrawStart',
+  'onDrawDrawstop',
+  'onDrawvertex',
+  'onDrawEditStart',
+  'onDrawEditMove',
+  'onDrawEditResize',
+  'onDrawEditvertex',
+  'onDrawEditStop',
+  'onDrawDeleteStart',
+  'onDrawDeleteStop'
+];
+
 @Component({
   selector: 'o-map',
   providers: [MapService, GeocodingService, MdIconRegistry],
-  inputs: [
-    ...OMapComponent.DEFAULT_INPUTS
-  ],
-  outputs: [
-    'onToggleWSLayerSelected',
-    'onToggleWSLayerVisibility',
-    'onToggleWSLayerInWS'
-  ],
+  inputs: OMapComponent.DEFAULT_INPUTS,
+  outputs: OMapComponent.DEFAULT_OUTPUTS,
   templateUrl: './o-map.component.html',
   styleUrls: ['./o-map.component.scss'],
   encapsulation: ViewEncapsulation.None
@@ -46,6 +68,7 @@ const DEFAULT_INPUTS = [
 export class OMapComponent extends OMapWSearch {
 
   public static DEFAULT_INPUTS = DEFAULT_INPUTS;
+  public static DEFAULT_OUTPUTS = DEFAULT_OUTPUTS;
 
   @ViewChild(OMarkerComponent) markerComponent: OMarkerComponent;
   @ViewChild('sidenav') sideNavCmp: MdSidenav;
@@ -78,6 +101,8 @@ export class OMapComponent extends OMapWSearch {
   protected mdTabGroupSubscription: Subscription;
   protected _waitForBuild: boolean = false;
   protected _searchControlButtonVisible: boolean = true;
+  protected crsComponent: OMapCrsComponent;
+
   constructor(
     protected elRef: ElementRef,
     protected injector: Injector
@@ -114,6 +139,14 @@ export class OMapComponent extends OMapWSearch {
     } else if (this.waitForBuild) {
       this.initialize();
     }
+    if (!this.mapService.map) {
+      // console.debug('Initializing map...');
+      this.configureMap();
+    }
+    if (this.drawDefaultControl && !this.drawControlComponent) {
+      this.configureDefaultDrawControl(this.mapService.getMap());
+    }
+    this.onMapReady().emit(this);
   }
 
   initialize() {
@@ -127,16 +160,15 @@ export class OMapComponent extends OMapWSearch {
     };
     this.searchControl = Util.parseBoolean(this.sSearchControl, true);
     this.isSearchControlButtonVisible = Util.parseBoolean(this.sSearchControlButtonVisible, true);
-    this.drawControl = Util.parseBoolean(this.sDrawControl, false);
+    this.drawDefaultControl = Util.parseBoolean(this.sDrawControl, false);
     this.isSidenavVisible = Util.parseBoolean(this.sLayerPanelVisible, false);
 
     this.baseLayerIds = Util.parseArray(this.sBaseLayerIds);
     this.mapService.configureBaseLayers(this.baseLayerIds);
+  }
 
-    if (!this.mapService.map) {
-      console.debug('Initializing map...');
-      this.configureMap();
-    }
+  registerCRSComponent(crsComp: OMapCrsComponent) {
+    this.crsComponent = crsComp;
   }
 
   registerTabGroupListener() {
@@ -198,19 +230,21 @@ export class OMapComponent extends OMapWSearch {
       zoom: this.zoom.current || 12,
       minZoom: this.zoom.min || 4,
       maxZoom: this.zoom.max || 19,
-      layers: this.mapService.baseLayers.getLayersArray(),
-      drawControl: false
+      layers: this.mapService.baseLayers.getLayersArray()
     };
+
+    if (this.crsComponent) {
+      let crsConf = this.crsComponent.getCRS();
+      if (crsConf) {
+        mapOptions['crs'] = crsConf;
+      }
+    }
 
     let map = this.mapService.getMap(this.mapId, mapOptions);
 
     L.control.scale().addTo(map);
     if (this.zoom.control) {
       L.control.zoom({ position: 'bottomright' }).addTo(map);
-    }
-
-    if (this.drawControl) {
-      //TODO  this.configureDrawControl(map);
     }
 
     this.onMapConfigured().emit(true);